Improve search results for "business plan"
There are 1,000 searches/month for "business plan". It seems likely that the
majority of them are after the 'Write a business plan' page, rather than the
Government Publications page. Because recommended-links/inside gov links are
artificially promoted to the top of search results, the useful result is being
pushed down.