Upgrade version of rummageable in use
This was using a very old version - 0.6.1.  The API of rummageable has
been updated since this version, so lots of changes subsequent to this
commit will be needed.

We need to upgrade rummageable because the old version didn't allow the
host to talk to to find rummager to be configured, and the name has
changed (from "search" to "rummager").  The new version does allow this
to be changed.