Use the deleted links csv to remove stale paths from Rummager
We have 67 base paths in Rummager representing Whitehall content served
from the /government/statistics/* base path. A bug in Whitehall means
that this content is served by whitehall-frontend at both
/government/statistics/* and /government/publications/*. The publishing
API is only aware of the 'publications' version of this content.

This commit uses the deleted-links CSV to remove the /statistics paths
from the search index, as the code in Whitehall and the paths found in
the publishing-api suggest that the /publications paths are the correct
ones.